write the logging section of the guide
