+++ ffi/bindings/dotnet/examples/master/Program.cs
@@ -6,7 +6,8 @@ using dnp3rs;
 class MainClass
 {
     // ANCHOR: logging_interface
-    class TestLogger : ILogger
+    // callback interface used to receive log messages
+    class ConsoleLogger : ILogger
     {
         public void OnMessage(LogLevel level, string message)
         {
@@ -147,13 +148,8 @@ class MainClass
 
     public static void Main(string[] args)
     {
-        MainAsync().GetAwaiter().GetResult();
-    }
-
-    private static async Task MainAsync()
-    {
-
         // ANCHOR: logging_init
+        // called once during program initialization
         Logging.Configure(
             new LoggingConfiguration
             {
@@ -163,171 +159,179 @@ class MainClass
                 TimeFormat = TimeFormat.System,
                 OutputFormat = LogOutputFormat.Text
             },
-            new TestLogger()
+            new ConsoleLogger()
         );
         // ANCHOR_END: logging_init
 
+        // ANCHOR: runtime_init
         using (var runtime = new Runtime(new RuntimeConfig { NumCoreThreads = 4 }))
+        // ANCHOR_END: runtime_init
         {
-            var master = runtime.AddMasterTcp(
-                new MasterConfiguration
-                {
-                    Address = 1,
-                    Level = DecodeLogLevel.ObjectValues,
-                    ReconnectionStrategy = new RetryStrategy
-                    {
-                        MinDelay = TimeSpan.FromMilliseconds(100),
-                        MaxDelay = TimeSpan.FromSeconds(5),
-                    },
-                    ReconnectionDelay = TimeSpan.Zero,
-                    ResponseTimeout = TimeSpan.FromSeconds(5),
-                    RxBufferSize = 2048,
-                    TxBufferSize = 2048,
-                    BubbleFramingErrors = false,
-                },
-                new EndpointList("127.0.0.1:20000"),
-                new TestListener()
-            );
+            MainAsync(runtime).GetAwaiter().GetResult();
+        }
+    }
 
-            var readHandler = new TestReadHandler();
-            var association = master.AddAssociation(
-                1024,
-                new AssociationConfiguration
-                {
-                    DisableUnsolClasses = new EventClasses
-                    {
-                        Class1 = true,
-                        Class2 = true,
-                        Class3 = true,
-                    },
-                    EnableUnsolClasses = new EventClasses
-                    {
-                        Class1 = true,
-                        Class2 = true,
-                        Class3 = true,
-                    },
-                    StartupIntegrityClasses = Classes.All(),
-                    AutoTimeSync = AutoTimeSync.Lan,
-                    AutoTasksRetryStrategy = new RetryStrategy
-                    {
-                        MinDelay = TimeSpan.FromSeconds(1),
-                        MaxDelay = TimeSpan.FromSeconds(5),
-                    },
-                    KeepAliveTimeout = TimeSpan.FromSeconds(60),
-                    AutoIntegrityScanOnBufferOverflow = true,
-                    EventScanOnEventsAvailable = new EventClasses
-                    {
-                        Class1 = false,
-                        Class2 = false,
-                        Class3 = false,
-                    },
-                },
-                new AssociationHandlers
-                {
-                    IntegrityHandler = readHandler,
-                    UnsolicitedHandler = readHandler,
-                    DefaultPollHandler = readHandler,
-                },
-                new TestTimeProvider()
-            );
+    private static async Task MainAsync(Runtime runtime)
+    {
 
-            var pollRequest = Request.ClassRequest(false, true, true, true);
-            var poll = association.AddPoll(pollRequest, TimeSpan.FromSeconds(5));
-
-            while (true)
+        var master = runtime.AddMasterTcp(
+            new MasterConfiguration
             {
-                switch (await GetInputAsync())
+                Address = 1,
+                Level = DecodeLogLevel.ObjectValues,
+                ReconnectionStrategy = new RetryStrategy
                 {
-                    case "x":
-                        return;
-                    case "dln":
-                        {
-                            master.SetDecodeLogLevel(DecodeLogLevel.Nothing);
-                            break;
-                        }
-                    case "dlv":
-                        {
-                            master.SetDecodeLogLevel(DecodeLogLevel.ObjectValues);
-                            break;
-                        }
-                    case "rao":
-                        {
-                            var request = new Request();
-                            request.AddAllObjectsHeader(Variation.Group40Var0);
-                            var result = await association.Read(request);
-                            Console.WriteLine($"Result: {result}");
-                            break;
-                        }
-                    case "rmo":
-                        {
-                            var request = new Request();
-                            request.AddAllObjectsHeader(Variation.Group10Var0);
-                            request.AddAllObjectsHeader(Variation.Group40Var0);
-                            var result = await association.Read(request);
-                            Console.WriteLine($"Result: {result}");
-                            break;
-                        }
-                    case "cmd":
-                        {
-                            var command = new Command();
-                            command.AddU16g12v1(3,
-                                new G12v1
-                                {
-                                    Code = new ControlCode
-                                    {
-                                        Tcc = TripCloseCode.Nul,
-                                        Clear = false,
-                                        Queue = false,
-                                        OpType = OpType.LatchOn,
-                                    },
-                                    Count = 1,
-                                    OnTime = 1000,
-                                    OffTime = 1000,
-                                }
-                            );
-                            var result = await association.Operate(CommandMode.SelectBeforeOperate, command);
-                            Console.WriteLine($"Result: {result}");
-                            break;
-                        }
-                    case "evt":
-                        {
-                            poll.Demand();
-                            break;
-                        }
-                    case "lts":
-                        {
-                            var result = await association.PerformTimeSync(TimeSyncMode.Lan);
-                            Console.WriteLine($"Result: {result}");
-                            break;
-                        }
-                    case "nts":
-                        {
-                            var result = await association.PerformTimeSync(TimeSyncMode.NonLan);
-                            Console.WriteLine($"Result: {result}");
-                            break;
-                        }
-                    case "crt":
-                        {
-                            var result = await association.ColdRestart();
-                            Console.WriteLine($"Result: {result}");
-                            break;
-                        }
-                    case "wrt":
-                        {
-                            var result = await association.WarmRestart();
-                            Console.WriteLine($"Result: {result}");
-                            break;
-                        }
-                    case "lsr":
-                        {
-                            var result = await association.CheckLinkStatus();
-                            Console.WriteLine($"Result: {result}");
-                            break;
-                        }
-                    default:
-                        Console.WriteLine("Unknown command");
+                    MinDelay = TimeSpan.FromMilliseconds(100),
+                    MaxDelay = TimeSpan.FromSeconds(5),
+                },
+                ReconnectionDelay = TimeSpan.Zero,
+                ResponseTimeout = TimeSpan.FromSeconds(5),
+                RxBufferSize = 2048,
+                TxBufferSize = 2048,
+                BubbleFramingErrors = false,
+            },
+            new EndpointList("127.0.0.1:20000"),
+            new TestListener()
+        );
+
+        var readHandler = new TestReadHandler();
+        var association = master.AddAssociation(
+            1024,
+            new AssociationConfiguration
+            {
+                DisableUnsolClasses = new EventClasses
+                {
+                    Class1 = true,
+                    Class2 = true,
+                    Class3 = true,
+                },
+                EnableUnsolClasses = new EventClasses
+                {
+                    Class1 = true,
+                    Class2 = true,
+                    Class3 = true,
+                },
+                StartupIntegrityClasses = Classes.All(),
+                AutoTimeSync = AutoTimeSync.Lan,
+                AutoTasksRetryStrategy = new RetryStrategy
+                {
+                    MinDelay = TimeSpan.FromSeconds(1),
+                    MaxDelay = TimeSpan.FromSeconds(5),
+                },
+                KeepAliveTimeout = TimeSpan.FromSeconds(60),
+                AutoIntegrityScanOnBufferOverflow = true,
+                EventScanOnEventsAvailable = new EventClasses
+                {
+                    Class1 = false,
+                    Class2 = false,
+                    Class3 = false,
+                },
+            },
+            new AssociationHandlers
+            {
+                IntegrityHandler = readHandler,
+                UnsolicitedHandler = readHandler,
+                DefaultPollHandler = readHandler,
+            },
+            new TestTimeProvider()
+        );
+
+        var pollRequest = Request.ClassRequest(false, true, true, true);
+        var poll = association.AddPoll(pollRequest, TimeSpan.FromSeconds(5));
+
+        while (true)
+        {
+            switch (await GetInputAsync())
+            {
+                case "x":
+                    return;
+                case "dln":
+                    {
+                        master.SetDecodeLogLevel(DecodeLogLevel.Nothing);
                         break;
-                }
+                    }
+                case "dlv":
+                    {
+                        master.SetDecodeLogLevel(DecodeLogLevel.ObjectValues);
+                        break;
+                    }
+                case "rao":
+                    {
+                        var request = new Request();
+                        request.AddAllObjectsHeader(Variation.Group40Var0);
+                        var result = await association.Read(request);
+                        Console.WriteLine($"Result: {result}");
+                        break;
+                    }
+                case "rmo":
+                    {
+                        var request = new Request();
+                        request.AddAllObjectsHeader(Variation.Group10Var0);
+                        request.AddAllObjectsHeader(Variation.Group40Var0);
+                        var result = await association.Read(request);
+                        Console.WriteLine($"Result: {result}");
+                        break;
+                    }
+                case "cmd":
+                    {
+                        var command = new Command();
+                        command.AddU16g12v1(3,
+                            new G12v1
+                            {
+                                Code = new ControlCode
+                                {
+                                    Tcc = TripCloseCode.Nul,
+                                    Clear = false,
+                                    Queue = false,
+                                    OpType = OpType.LatchOn,
+                                },
+                                Count = 1,
+                                OnTime = 1000,
+                                OffTime = 1000,
+                            }
+                        );
+                        var result = await association.Operate(CommandMode.SelectBeforeOperate, command);
+                        Console.WriteLine($"Result: {result}");
+                        break;
+                    }
+                case "evt":
+                    {
+                        poll.Demand();
+                        break;
+                    }
+                case "lts":
+                    {
+                        var result = await association.PerformTimeSync(TimeSyncMode.Lan);
+                        Console.WriteLine($"Result: {result}");
+                        break;
+                    }
+                case "nts":
+                    {
+                        var result = await association.PerformTimeSync(TimeSyncMode.NonLan);
+                        Console.WriteLine($"Result: {result}");
+                        break;
+                    }
+                case "crt":
+                    {
+                        var result = await association.ColdRestart();
+                        Console.WriteLine($"Result: {result}");
+                        break;
+                    }
+                case "wrt":
+                    {
+                        var result = await association.WarmRestart();
+                        Console.WriteLine($"Result: {result}");
+                        break;
+                    }
+                case "lsr":
+                    {
+                        var result = await association.CheckLinkStatus();
+                        Console.WriteLine($"Result: {result}");
+                        break;
+                    }
+                default:
+                    Console.WriteLine("Unknown command");
+                    break;
             }
         }
     }
