+++ guide/sidebars.js
@@ -2,6 +2,6 @@ module.exports = {
   someSidebar: {
     About: ['about/dnp3', 'about/library'],
     API: ['api/logging'],
-    Features: ['code', 'table'],
+    Features: ['table'],
   },
 };
